jetson_nano/requirements.txt
Jetson.GPIO==2.0.17
matplotlib==2.1.1
numpy==1.19.5
Pillow==8.3.1
pycuda==2022.1
scipy==0.19.1
tensorrt==8.0.1.6
tqdm==4.62.0

nanocamera==0.1.5